fix: 修复默认图片会被删除bug

+++ apps/users/api.py
@@ -5,7 +5,7 @@ from django.urls import reverse
 from rest_framework import mixins
 from rest_framework import permissions
 from rest_framework import viewsets
-from rest_framework.decorators import list_route
+from rest_framework.decorators import action
 from rest_framework.exceptions import ValidationError
 from rest_framework.response import Response
 from rest_framework.status import HTTP_200_OK, HTTP_400_BAD_REQUEST
@@ -25,7 +25,6 @@ class RecordIPViewSet(CacheResponseMixin,
     permission_classes = (permissions.AllowAny,)
 
 
-
 class UserViewSet(CacheResponseMixin, viewsets.ModelViewSet):
     queryset = User.objects.all()
     serializer_class = UserSerializers
@@ -42,11 +41,12 @@ class UserViewSet(CacheResponseMixin, viewsets.ModelViewSet):
 
         new_instance = serializer.save()
 
-        if old_image_name != new_instance.image.name and os.path.exists(old_image_path):
+        if old_image_name != new_instance.image.name \
+                and old_image_name != 'image/default.jpg' \
+                and os.path.exists(old_image_path):
             os.remove(old_image_path)
 
-
-    @list_route(methods=['POST'], permission_classes=(permissions.AllowAny,))
+    @action(methods=['POST'], permission_classes=(permissions.AllowAny,), detail=False)
     def check_user_register(self, request):
         username = request.data.get('username', '')
         email = request.data.get('email', '')
@@ -60,7 +60,7 @@ class UserViewSet(CacheResponseMixin, viewsets.ModelViewSet):
                 return Response(data=False, status=HTTP_200_OK)
         return Response(data=True, status=HTTP_200_OK)
 
-    @list_route(methods=['POST',], permission_classes=(permissions.AllowAny,))
+    @action(methods=['POST', ], permission_classes=(permissions.AllowAny,), detail=False)
     def send_eamil(self, request):
         email = request.data.get('email', '')
         if not email:
@@ -84,7 +84,7 @@ class UserViewSet(CacheResponseMixin, viewsets.ModelViewSet):
         custom_send_mail(title, msg, email_from, email_to)
         return Response(data=True, status=HTTP_200_OK)
 
-    @list_route(methods=['POST', ])
+    @action(methods=['POST', ], detail=False)
     def reset_pwd(self, request):
         pwd = request.data.get('password', '')
         pwd2 = request.data.get('password2', '')
